aerowise_llm_agent/requirements.txt
# Framework Agents IA
agno>=0.1.0

anthropic>=0.1.0

# LLM Providers
openai>=1.0.0

# Data manipulation
pandas>=2.0.0
pyarrow>=14.0.0  # Pour lire les .parquet

# Validation de données
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Utils
python-dateutil>=2.8.0
